Major (Hopeful) Combat Breakthrough
Potential breakthrough on combat control. Changed runCombat function to auto process computer turn as before, got it to successfully wait for a button click on player turn/assign the entire combat processing to a timer so that it doesn't freeze out the entire application, but it still waits until the player choice is made. All updates still currently print statements to the console for testing, feel free to give it a shot (a bit annoying, but you can just jump between the game window and Eclipse).

Tweaked a few other things in relation to character positioning, and things that related to combat stuff.
Additionally altered the .gitignore pushes SHOULD no longer overwrite our individual library information for FX.