Functional Combat!!
Finished tweaking combatFlow stuff, you can now play trough a fully functional round, and it finishes processing. If you go back to the main menu and load back in, it will work again as well.

Fixed a data leak in determineTurnOrder that was messing with the turn count.

Fixed some logic for the function to check if either team is dead as well.